feat: add error handling to resource not found in endpoint findEmployeeById  and deleteEmployee

+++ src/main/java/com/exemple/employee_register/employee/Services/EmployeeService.java
@@ -2,10 +2,10 @@ package com.exemple.employee_register.employee.Services;
 
 import com.exemple.employee_register.employee.Models.EmployeeModel;
 import com.exemple.employee_register.employee.repositories.IEmployeeRepository;
+import com.exemple.employee_register.exceptions.NotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Service
@@ -22,15 +22,19 @@ public class EmployeeService {
         return this.employeeRepository.findAll();
     }
 
-    public Optional<EmployeeModel> findById(UUID id){
-        return  this.employeeRepository.findById(id);
+    public EmployeeModel findById(UUID id) {
+        return this.employeeRepository.findById(id).orElseThrow(() -> new NotFoundException("Employee not found"));
     }
 
-    public EmployeeModel create(EmployeeModel employee){
-        return  this.employeeRepository.save(employee);
+    public EmployeeModel create(EmployeeModel employee) {
+        return this.employeeRepository.save(employee);
     }
 
-    public void delete(UUID id){
-         this.employeeRepository.deleteById(id);
+    public void delete(UUID id) {
+
+        if (!employeeRepository.existsById(id)) {
+            throw new NotFoundException("Employee not found");
+        }
+        this.employeeRepository.deleteById(id);
     }
 }
